user docs: Remove allow-anyone-to-join-without-an-invitation.
The information here was recently added to manage-who-can-join-and-invite.

Arguably this is one we should save, since it is a distinctive feature not
offered by all of our competitors, and it gets some additional visibility by
being in the left sidebar. The model of having multiple things in the
sidebar pointing to the same article is getting messy though, and as our
feature count increases the cost of having stuff in the left sidebar is
increasing as well.