export: Fix error messages for stream list mismatches.
The previous error messages for this were written for a tool only to
be used by a couple people, and didn't make clear what potential
causes were.  Tweak these to provide greater clarity about what's
going on.

The main cause of these errors appearing in practice was fixed in
7ea5987e5d061ef22bebef5fdb1d6bba0010e8d1, but nothing strongly
prevents a similar issue from being introduced in the future.

Fixes #10078.